reuse existing alternative titles function
